requirements.txt
# API
fastapi>=0.116
uvicorn[standard]>=0.30
pydantic>=2.7,<3
python-multipart>=0.0.9

# ML/CV
numpy>=2,<2.3           
opencv-python-headless>=4.12.0.88
Pillow>=10.4
ultralytics>=8.3
albumentations>=1.4
timm>=1.0
scikit-image>=0.24       #
scikit-learn>=1.4.2     
pycocotools>=2.0.8

# DL
torch>=2.2             
torchvision>=0.17

# HF / OVD
huggingface-hub>=0.23
transformers>=4.42
ftfy>=6.2
regex>=2024.5.15

# Конфиги / утилиты
hydra-core>=1.3
omegaconf>=2.3
tqdm>=4.66
loguru>=0.7
rich>=13.7

# Экспорт / рантайм
onnx>=1.16
onnxruntime; sys_platform == "darwin"     # CPU/Metal на macOS
# onnxruntime-gpu; sys_platform == "linux"  # Linux+CUDA

# SAHI / slicing
sahi>=0.11

# GroundingDINO (требует PyTorch)
groundingdino-py>=0.4
